OCT-111: moved GetProductUuidFromIdentifierQuery
