+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/Catalog/Product/GetProductUuidFromIdentifierQueryTest.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence;
+namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence\Catalog\Product;
 
-use Akeneo\Catalogs\Infrastructure\Persistence\GetProductUuidFromIdentifierQuery;
+use Akeneo\Catalogs\Infrastructure\Persistence\Catalog\Product\GetProductUuidFromIdentifierQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @covers \Akeneo\Catalogs\Infrastructure\Persistence\GetProductUuidFromIdentifierQuery
+ * @covers \Akeneo\Catalogs\Infrastructure\Persistence\Catalog\Product\GetProductUuidFromIdentifierQuery
  */
 class GetProductUuidFromIdentifierQueryTest extends IntegrationTestCase
 {
